BUG: use INT rather than LONG for label map type
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12870
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12870 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/OtsuThresholdSegmentation.cxx
@@ -51,7 +51,7 @@ int main( int argc, char * argv[] )
 // Image Types
   typedef  long  InputPixelType;
   typedef  long  InternalPixelType;
-  typedef  long  OutputPixelType;
+  typedef  int   OutputPixelType;
 
   typedef itk::Image< InputPixelType,  3 >   InputImageType;
   typedef itk::Image< InternalPixelType, 3 >   InternalImageType;
